Don't assume in plot.fprojsel() that the getfullname() function exists.
To plot nice-looking labels, a function called getfullname() can be defined
in the workspace in such a way that variable names are matched to full names
with the following signature:

   getfullname <- function(name) { ... }

where the body of the function can be any desired transformation of the
variable names.

This change allows the code to run also if such function doesn't exist.