Simplify the implementation of sample.stan().
This should make it more obvious that training and test data are meant to
be the same if there is no cross-validation. This is due to the fact that
the Stan model expects both a training and a test set, and predicted values
are computed only for the test observations.