Rewrite the package by using the facilities provided by rstantools.
This approach precompiles the stan models when the package is installed, but
avoids the issues related to passing around the stan model files.

Therefore, the choice between a base model and a model with biomarkers can
be made based on whether the biomarkers are specified (in which case they
are penalised with hierarchical shrinkage), or they are an empty vector or
NULL (in which case a covariates-only model is used).

The choice of logistic regression is now done via an option, although now we
do not yet check that this is consistent with the outcome variable given.