Fix bug in the assignment of readable variable names.
The columns in the object produced by Stan are ordered according to their
declaration in the model. Therefore, after beta_u there is not beta_p as the
previous code implied, but z, which is generally quite close to beta_p (as
beta_p is a transformation of z), but they are not identical.

The bug affected only the betas and coefficients reported by sample.stan().
It didn't affect the predictions, the fitted value, nor the projective
selection.