Add a loo() method for hsstan objects.
The double @export statement is necessary because otherwise the loo()
function wouldn't get properly exported, likely due to the fact that we
import an S3 method from rstan which in turn imports the S3 method from loo,
and neither package is in hsstan's Depends section. This solution is the one
adopted by brms, which likely faced the same issue.

The @aliases statement is necessary because otherwise ?loo would only offer
loo::loo or rstan::loo as options for documentation.