Don't validate newdata twice in log_lik().
Validation of newdata already occurs inside posterior_linpred(). If newdata
contains a factor variable, after the first validation its name changes
(say, from sex to sex.F), and that column is dropped during the second
validation as sex.F is not a name present in the variable names stored in
model.terms.