Add the kfold() function to cross-validate a fitted hsstan object.
This provides a more flexible way to cross-validate a model, which will
simplify the code for hsstan() and the Stan models. The result produced is
compatible with loo::loo_compare() for comparing models.

Results are slightly different from before when cross-validating because the
QR decomposition now effectively happens at the fold level (which is more
correct), while before it was done before the training/test split.