Drop the internal horseshoe parameters from the stanfit object by default.
This adds the keep.hs.pars option to control whether the internal horseshoe
parameters should be kept or not. By default they are dropped, so that the
resulting stanfit object is much smaller.

For a model with 50 predictors and 300 observations, the size of the hsstan
object goes from 7.7M to 2.9M, and a 5-fold model from 41.5M to 17.8M.