Fix bug in handling penalized variables mentioned in the unpenalized model.
If a variable appeared in both the unpenalized model and in the vector of
penalized predictors, the penalization was potentially applied to the wrong
variables, as those predictors were not removed from the unpenalized set.

For example, in this model:

   hsstan(df, y ~ X1 + X2 + X3, "X2")

the unpenalized set was considered to be X1, X2, X3 and the penalty was
applied to X3, being the last element in the model matrix. With this change,
instead, the unpenalized set becomes X1, X3 and the design matrix lists X2
last, so that the penalized variable is now X2 as desired.