Implement matrix multiplication with RcppEigen.
This replaces the uses of %*% and crossprod() when they are applied to two
different matrices: in such cases it is often twice as fast, and even more
so for very large matrices.

The uses of crossprod() applied to the same matrix are not removed as, being
able to exploit symmetry in the result, crossprod() is much faster than the
version based on RcppEigen, especially if both arguments are explicitly used.